add h8 start
